Work in Progress GameEngine project. Fixing mix signed/unsigned and VS unsecure syntax

+++ Generals/Code/GameEngine/Include/Common/AsciiString.h
@@ -156,7 +156,7 @@ public:
 	/**
 		Return the length, in characters (not bytes!), of the string.
 	*/
-	int getLength() const;
+	size_t getLength() const;
 	/**
 		Return true iff the length of the string is zero. Equivalent
 		to (getLength() == 0) but slightly more efficient.
@@ -363,7 +363,7 @@ inline AsciiString::~AsciiString()
 }
 
 // -----------------------------------------------------
-inline int AsciiString::getLength() const
+inline size_t AsciiString::getLength() const
 {
 	validate();
 	return m_data ? strlen(peek()) : 0;
@@ -387,7 +387,7 @@ inline void AsciiString::clear()
 // -----------------------------------------------------
 inline char AsciiString::getCharAt(int index) const
 {
-	DEBUG_ASSERTCRASH(index >= 0 && index < getLength(), ("bad index in getCharAt"));
+	DEBUG_ASSERTCRASH(index >= 0 && (size_t)index < getLength(), ("bad index in getCharAt"));
 	validate();
 	return m_data ? peek()[index] : 0;
 }
